Flask/flask_template/requirements.txt
click==7.1.2
dash==1.14.0
dash-bootstrap-components==0.10.7
email-validator==1.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
JayDeBeApi==1.2.3
numpy==1.19.1
pandas==1.1.0
plotly==4.9.0
PyJWT==1.7.1
python-dotenv==0.15.0
sqlalchemy==1.3.20
werkzeug==1.0.1
wtforms==2.3.3
